Rewrite in Bash
Since Bash is more common in the Nix ecosystem, I figured it was a waste
of closure size to have this script depend on Zsh.

I also took this change as an opportunity to make sure the quoting was
rock-solid. I had assumes that Zsh used TCL-like quoting by default,
however a recent bug indicated that this wasn't actually the case.